requirements.txt
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.5
blinker==1.4
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==25.1.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
et_xmlfile==2.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gspread-dataframe==4.0.0
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==6.0.1
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
openpyxl==3.1.5
pandas==2.3.2
pyaes==1.6.1
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
soupsieve==2.8
systemd-python==234
Telethon==1.40.0
Twisted==22.1.0
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
